services/voice-search/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
httpx==0.25.1
python-dotenv==1.0.0
speechrecognition==3.10.0
pydub==0.25.1
pymongo==4.6.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
numpy==1.26.1
scikit-learn==1.3.2
transformers==4.35.0
torch==2.1.0
soundfile==0.12.1
librosa==0.10.1
